pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vijay-ai-codegen"
version = "0.1.0"
description = "Python code generator & optimizer using Euri AI"
readme = "README.md"
authors = [{ name = "Your Name" }]
requires-python = ">=3.10"
dependencies = [
  "euriai>=0.1.0",
  "typer>=0.12.0",
  "rich>=13.7.0",
  "pydantic>=2.5",
  "python-dotenv>=1.0.1",
  "streamlit>=1.37.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "pytest-benchmark>=4.0.0",
  "black>=24.3.0",
  "ruff>=0.4.0",
  "mypy>=1.8.0",
]

[project.scripts]
vijay-codegen = "euri_codegen.cli:app"

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.10"
strict = false

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
"euri_codegen.catalog" = ["*.json"]
